Use coreutils for consistent `ls` color flags
Some directories that use direnv will pull in coreutils, which
includes the GNU version of `ls` that accepts different flags from the
BSD version included in macOS. This leads to some inconsistencies where
`ls` will return colored output in some directories but not others.

This PR adds coreutils to the main package set, so that we can just use
the GNU versions everywhere.