Use fork of nixpkgs-mozilla (exclude rustdoc)
Installing Rust with nixpkgs-mozilla is painfully slow due to the
installation of docs. There's no option to exclude the docs with the
standard installation, but there is an open PR for exposing a flag for
it. It doesn't look like the PR will be merged anytime soon, so updating
locally to use a fork of the repo that has the fix.