Remove nix-managed watchman
The version in nixpkgs is very out of date and doesn't work with modern
projects. I tried adding a derivation that pulls it from github, which
also seems to have issues (it hangs forever for some unknown reason).
So unfortunately I'm just going to remove it and use the homebrew
packaged version.