Replace telescope-nvim with fzf-lua
My main use-case for fzf-vim originally was fuzzy finding across all
files in the directory. It can be approximated with `grep_string` with
an empty search string, but may be slow since it doesn't do caching
like fzf-vim does.

Relevant comment:
https://github.com/nvim-telescope/telescope.nvim/issues/564#issuecomment-786876496

Deciding to try out fzf-lua for now, since a lot of its functionality is
similar to what was in fzf-vim.